requirements.txt
Flask==1.1.1
gunicorn==19.9.0
line-bot-sdk==1.14.0
pymsteams==0.1.13
requests==2.24.0
urllib3==1.25.10